nftProject/requirements.txt
fastapi
uvicorn
requests
pillow
opencv-python-headless
numpy
python-multipart
python-dotenv